hratuga-api/django/requirements.txt
argcomplete==1.12.2
asgiref==3.3.4
boto3==1.17.56
botocore==1.20.56
certifi==2020.12.5
cffi==1.14.5
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-keycloak @ git+https://github.com/Peter-Slump/django-keycloak.git@5211fba26ae4216b4344309eb4a67da9745ada05
django-rest-auth==0.9.5
django-storages==1.6.6
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-api-tracking==1.7.8
drf-keycloak-auth==0.0.1.dev12
durationpy==0.5
ecdsa==0.17.0
future==0.18.2
hjson==3.0.2
idna==2.10
jmespath==0.10.0
kappa==0.6.0
oauthlib==3.1.0
pep517==0.10.0
Pillow==8.1.2
pip-tools==6.1.0
placebo==0.9.0
progressbar2==3.53.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
python-dateutil==2.6.1
python-jose==3.3.0
python-keycloak==0.26.1
python-keycloak-client==0.2.3
python-memcached==1.59
python-slugify==4.0.1
python-utils==2.5.6
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
tqdm==4.59.0
troposphere==2.6.3
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.50.0
